Read the site's colours again from the brief
The palette was only ever written by the onboarding wizard, and the wizard
refuses a launched project — so every project that existed before it could never
be offered a suggestion at all. A button on the brief closes that.

**The colours, and nothing else.** Re-running SiteAnalyst would overwrite the
whole of `site_analysis`: the name, the description, the audiences, the seed
keywords, every one of which the operator corrected by hand in the wizard and
none of which they asked to have re-guessed. A button labelled "read my site"
that silently replaced a month of corrections is the worst kind of destructive,
because nothing on screen would show what it had taken. This touches one key.

Still a suggestion when it lands: it refreshes what the swatches offer, and
applying them stays a click.

Two bugs fixed along the way, both found by running it rather than by reading it:

- SitePalette divided by zero on any page containing pure black. PHP's `/`
  returns an int when both operands are ints and the division is exact, so
  `0x00 / 255` is `int(0)` while every other channel is a float — and the strict
  `=== 0.0` achromatic guard missed it. No fixture caught this: the 4-bit
  quantisation turns 255 into 240, so only genuine black reaches it, which every
  real screenshot has and no generated page did.
- home/index.tsx never rendered its "Waiting for you" band. `Deferred` supplies
  the fallback and does not hand the prop to its children, so `waiting` was
  permanently undefined and the skeleton showed forever — indistinguishable from
  a slow request that never lands.

+++ app/app/Support/Brand/SitePalette.php
@@ -214,9 +214,18 @@ final readonly class SitePalette
     {
         $hex = ltrim($hex, '#');
 
-        $r = ((int) hexdec(substr($hex, 0, 2))) / 255;
-        $g = ((int) hexdec(substr($hex, 2, 2))) / 255;
-        $b = ((int) hexdec(substr($hex, 4, 2))) / 255;
+        // Cast to float *before* dividing, and it is not decoration. PHP's `/`
+        // returns an int when both operands are ints and the division is exact,
+        // so `0x00 / 255` is `int(0)` and `0xff / 255` is `int(1)` — while every
+        // other channel value comes back a float. On a pure black or pure white
+        // pixel the delta below was therefore `int(0)`, the strict `=== 0.0`
+        // guard did not match it, and the achromatic case fell through to a
+        // division by that same zero. It never fired in tests because the
+        // 4-bit quantisation turns 255 into 240, so only genuine black reaches
+        // it — which every real screenshot has and no generated fixture did.
+        $r = ((float) hexdec(substr($hex, 0, 2))) / 255;
+        $g = ((float) hexdec(substr($hex, 2, 2))) / 255;
+        $b = ((float) hexdec(substr($hex, 4, 2))) / 255;
 
         $max = max($r, $g, $b);
         $min = min($r, $g, $b);
